+++ 05-ScenceManager/MapBackground.h
@@ -25,7 +25,7 @@ class MapBackground
 
 
 public:
-	MapBackground::MapBackground(int mapcolumns, int maprows, int tileheight, int tilewidth, int tilesetrows, int tilesetcolumns);
+	MapBackground::MapBackground(int mapcolumns, int maprows, float tileheight, float tilewidth, int tilesetrows, int tilesetcolumns);
 	
 	void LoadMatrix(LPCWSTR MatrixfilePath);
 	void DrawTile(int tileID, LPDIRECT3DTEXTURE9 tileset, int i, int j);
